fix: Correctly format logged errors during arg parsing
Right now, we don't log any errors at that time,
but the old code would print inconsistent results
which was a bit annoying. There is technically a
circular dependency, since the verbosity level
is determined during argument parsing, and it
is needed to determine the correct logging level.

Ignoring that as an edge case.